requirements.txt
pandas
openpyxl
sentence-transformers
faiss-cpu
fastapi
pydantic
uvicorn
openai
torch